requirements.txt
jsonrpclib-pelix
newspaper3k
numpy
pandas
pika
pymongo
python-dateutil
redis
requests
sklearn
scipy
tensorflow
watchdog